+++ dist-extension/service-worker-loader.js
@@ -1,1 +1,1 @@
-import './assets/background.ts-I204vJgW.js';
+import './assets/background.ts-IOZfUn4Z.js';